api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-ai
python-multipart
sse-starlette
httpx
pymupdf
python-dotenv
